Adding test_plotting and moving main plotting tests here
